requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.2
Django==5.2.1
django-allauth==65.8.0
gunicorn==23.0.0
idna==3.10
numpy==2.2.6
oauthlib==3.2.2
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
qrcode==8.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
